Refactor mongo game
